+++ CHANGELOG.md
@@ -4,4 +4,4 @@
 * satella.coding.Context are considered experimental
 * added a way to register an object to be cleaned up via MemoryPressureManager
 * fixed get_own_cpu_usage() to work on Windows
-
+* added __len__ to Optional